refactor: remove `whatwg-fetch` and `fetch-intercept` dependencies
Closes #64
