+++ README.md
@@ -16,7 +16,7 @@ refactoring the application to allow for two-way transfers.
 
 The tool is step-based, so you start by selecting the action you'd like to perform, then you enter your API keys, and so on. There are help details associated with each step on the [website](https://transfermyti.me).
 
-## Production Requirements
+You'll need the following to use the application:
 
 - Toggl account with API key
 - Clockify account with API key
@@ -61,11 +61,11 @@ section.
 4. Set `TMT_LOCAL_API_<TOOL>_EMPTY` to `true` in the `.env` file if you want either of the mock APIs to return empty records (useful for testing transfer)
 5. Run `pnpm start:server` (no data will be transferred from Toggl to Clockify)
 6. Run `pnpm start:web` to start the web application
-7. Navigate to `http://localhost:9008`
+7. Navigate to `http://localhost:8080`
 
 #### Development Without the Mock Server
 
 1. Install dependencies with `pnpm install`
 2. Rename `.env.example` to `.env`
 3. Run `pnpm start:web -- --env TMT_USE_LOCAL_API=false`
-4. Navigate to `http://localhost:9008`
+4. Navigate to `http://localhost:8080`
