+++ package.json
@@ -13,7 +13,7 @@
     "clockify:write": "node scripts/run.mjs clockify-write",
     "lint": "eslint --ext \".js,.mjs,.ts,.tsx\" \"{scripts,src}/**/*.{js,mjs,ts,tsx}\"",
     "prettier": "prettier --write \"{scripts,src,server}/**/*.{js,ts,tsx,css,html}\"",
-    "start:web": "webpack serve --port 9008 --mode development",
+    "start:web": "webpack serve --port 8080 --mode development",
     "start:server": "node scripts/run.mjs serve",
     "stylelint": "stylelint \"src/**/*.{css,html,tsx}\"",
     "test": "NODE_ENV=test node scripts/run.mjs tests --config jest.config.json",
@@ -25,7 +25,6 @@
     "connected-react-router": "^6.9.2",
     "cuid": "^2.1.8",
     "date-fns": "^2.28.0",
-    "fetch-intercept": "^2.4.0",
     "history": "4.10.1",
     "qs": "^6.10.3",
     "ramda": "^0.28.0",
@@ -38,8 +37,7 @@
     "redux": "^4.1.2",
     "redux-saga": "^1.1.3",
     "reselect": "^4.1.5",
-    "typesafe-actions": "^5.1.0",
-    "whatwg-fetch": "^3.6.2"
+    "typesafe-actions": "^5.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.17.8",
