+++ src/declarations.d.ts
@@ -1,5 +1,3 @@
-declare function fetch(input: RequestInfo, init?: RequestInit): any;
-
 declare type AnyFixMe = any;
 
 // Used to avoid the no-explicit-any ESLint rule explicitly. In some cases, it
