+++ src/index.tsx
@@ -1,4 +1,3 @@
-import "whatwg-fetch";
 import { createBrowserHistory } from "history";
 import React from "react";
 import { render } from "react-dom";
@@ -8,12 +7,9 @@ import "./styles.css";
 
 import { configureStore } from "~/redux/configureStore";
 import Routes from "~/Routes";
-import { initInterceptor } from "~/utilities/httpInterceptor";
 
 const history = createBrowserHistory();
 const store = configureStore(history);
-
-initInterceptor(store);
 
 render(
   <Provider store={store}>
